Rename swankyot
Our implementation of OT already differs from the one in swanky
and we intend to make more changes. To not cause confusion, we rename
our OT implementation.
