+++ src/lib.rs
@@ -97,8 +97,8 @@ mod faand;
 mod fpre;
 mod garble;
 mod ot;
+mod ot_core;
 mod rand_compat;
-mod swankyot;
 // TODO remove this once OT implementations are refactored and we know
 // what parts we need and which not
 #[allow(dead_code)]
