+++ src/ot_core/chou_orlandi.rs
@@ -20,8 +20,8 @@ use crate::{
     block::Block,
     channel::{recv_vec_from, send_to, Channel},
     faand::Error,
+    ot_core::{Malicious, Receiver as OtReceiver, SemiHonest, Sender as OtSender},
     rand_compat::RngCompat,
-    swankyot::{Malicious, Receiver as OtReceiver, SemiHonest, Sender as OtSender},
 };
 
 use curve25519_dalek::{
